backend/requirements.txt
appdirs==1.4.4
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
psycopg2-binary==2.9.9
PyExecJS==1.5.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pyTwistyScrambler==1.3
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
wsproto==1.2.0
